requirements.txt
# Core
pandas>=2.0.0
numpy>=1.24.0
duckdb>=0.9.0

# ML
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
joblib>=1.3.0

# MCP Server
mcp>=1.0.0

# Utils
python-dotenv>=1.0.0
